Pass unknown options to latexdiff
This only works for options of type --longoption=something, and not for
options of type -shortoption something. But since every short option in
latexdiff has a corresponding long option, this should still allows the
user to pass every possible options.

For instance to call
    latexdiff -c FLOATENV=foo,ARRENV=bar
use
    git-latexdiff --config="FLOATENV=foo,ARRENV=bar"

The implementation has the side effect of allowing
    git-latexdiff -- OLD
to be similar to
    git-latexdiff OLD --
Moreover it uses an array to store the options to latexdiff so we switch
from sh to bash.

Signed-off-by: Damien Robert <damien.olivier.robert+git@gmail.com>