Expand the array $latexdiffopt
In bash (unlike zsh), using $array only expand to the first element of the
array. To expand to the whole array, we have to use ${array[@]}. Since we
did not expand $latexdiffopt, this meant that only the last option would be
passed to latexdiff. Now we can pass several options (which was the point
of using an array in the implementation in the first place).

Signed-off-by: Damien Robert <damien.olivier.robert+git@gmail.com>