Allow using latexpand instead of latexdiff --flatten
"latexdiff --flatten" is not very robust. It doesn't expand bibliography,
and doesn't seem to work when the \begin{document} is inside an \input-ed
file.

Instead, use latexpand before latexdiff, which should solve these issues.

Works with --latexpand <options>, but has to be last in the series of
options. Shortcut --bbl to include the bibliography with the same name as
the project file.

Commit message improved by: Matthieu Moy <Matthieu.Moy@grenoble-inp.fr>