New option --ignore-latex-errors
This option causes git-latexdiff to be permissive about errors generated
by running LaTeX on the difference file, so long as a PDF file is
produced.  I have found this to be necessary for documents with complex
math markup and that use the `amsmath` package.  The difference file
tends to generate many "Multiple \label's" errors, although these are
harmless and do not affect the output PDF file.