Support Emacs Org Mode based diffs
We load ~/.emacs as --batch implies -q
but some customization might be expected.

For the same reason something like --tmpdirprefix .
is necessary for .dir-locals.el to work.

Finally, if on Windows and using Chocolatey,
real emacs and not shim is neccesary
as shim doesn't wait for process completion.

path %ChocolateyInstall%\lib\emacs64\tools\emacs\bin;%PATH%