don't call "git archive" with empty pathspec
When called from the top of the repository, $git_prefix is empty, and
setting checkoutroot to $git_prefix implies calling git with an empty
string as pathspec. This behavior was accepted as meaning "all path" in
old Git, but now triggers a warning and will soon be completely
forbidden.

Fixes #28.